cron

+++ scripts/cron.php
@@ -181,21 +181,25 @@ if ($taskcount > 0) {
         ))
       );
       $tunneljson = curl_exec($postchtunnels);
-      print_r($tunneljson);
-      echo ("Deleting existing tunnels\n");
-      fwrite($logfile, "\n" . date("Y-m-d h:i:sa") . " - Deleting existing tunnels.");
       $tunnelarr = json_decode($tunneljson);
-      $tunnelcount = count($tunnelarr->tunnellist);
-      fwrite($logfile, "\n" . date("Y-m-d h:i:sa") . " - Syncronizing $tunnelcount tunnels.");
-      echo ("Syncronizing $tunnelcount tunnels\n");
-      mysqli_query($dbconn, "DELETE FROM tunnels");
-      for ($y = 0; $y <= $tunnelcount - 1; $y++) {
-        fwrite($logfile, "\n" . date("Y-m-d h:i:sa") . " - Adding " . $tunnelarr->tunnellist[$y]->application . " tunnel. " . $tunnelarr->tunnellist[$y]->remotehost . ":" . $tunnelarr->tunnellist[$y]->remoteport . " to " . $tunnelarr->tunnellist[$y]->tunnelport . ".");
-        echo ("Adding " . $tunnelarr->tunnellist[$y]->application . " tunnel. " . $tunnelarr->tunnellist[$y]->remotehost . ":" . $tunnelarr->tunnellist[$y]->remoteport . " to " . $tunnelarr->tunnellist[$y]->tunnelport . "\n");
-        mysqli_query($dbconn, "INSERT INTO tunnels (`tunnelname`, `tunnelport`, `localproto`, `localhost`, `localport`) VALUES ('" . $tunnelarr->tunnellist[$y]->application . "', '" . $tunnelarr->tunnellist[$y]->tunnelport . "', '" . $tunnelarr->tunnellist[$y]->remoteproto . "', '" . $tunnelarr->tunnellist[$y]->remotehost . "', '" . $tunnelarr->tunnellist[$y]->remoteport . "')");
+      print_r($tunneljson);
+      if ($tunnelarr->tunnels > 0) {
+        echo ("Deleting existing tunnels\n");
+        fwrite($logfile, "\n" . date("Y-m-d h:i:sa") . " - Deleting existing tunnels.");
+        $tunnelcount = count($tunnelarr->tunnellist);
+        fwrite($logfile, "\n" . date("Y-m-d h:i:sa") . " - Syncronizing $tunnelcount tunnels.");
+        echo ("Syncronizing $tunnelcount tunnels\n");
+        mysqli_query($dbconn, "DELETE FROM tunnels");
+        for ($y = 0; $y <= $tunnelcount - 1; $y++) {
+          fwrite($logfile, "\n" . date("Y-m-d h:i:sa") . " - Adding " . $tunnelarr->tunnellist[$y]->application . " tunnel. " . $tunnelarr->tunnellist[$y]->remotehost . ":" . $tunnelarr->tunnellist[$y]->remoteport . " to " . $tunnelarr->tunnellist[$y]->tunnelport . ".");
+          echo ("Adding " . $tunnelarr->tunnellist[$y]->application . " tunnel. " . $tunnelarr->tunnellist[$y]->remotehost . ":" . $tunnelarr->tunnellist[$y]->remoteport . " to " . $tunnelarr->tunnellist[$y]->tunnelport . "\n");
+          mysqli_query($dbconn, "INSERT INTO tunnels (`tunnelname`, `tunnelport`, `localproto`, `localhost`, `localport`) VALUES ('" . $tunnelarr->tunnellist[$y]->application . "', '" . $tunnelarr->tunnellist[$y]->tunnelport . "', '" . $tunnelarr->tunnellist[$y]->remoteproto . "', '" . $tunnelarr->tunnellist[$y]->remotehost . "', '" . $tunnelarr->tunnellist[$y]->remoteport . "')");
+        }
+        fwrite($logfile, "\n" . date("Y-m-d h:i:sa") . " - Restarting tunnel service.");
+        exec('service callmigrate-tunnel restart');
+      } else {
+        echo ("No tunnels to sync!\n");
       }
-      fwrite($logfile, "\n" . date("Y-m-d h:i:sa") . " - Restarting tunnel service.");
-      exec('service callmigrate-tunnel restart');
       echo("DELETE FROM `tasks` WHERE `pkid` = " . $rowtasks["pkid"]);
       fwrite($logfile,"DELETE FROM `tasks` WHERE `pkid` = " . $rowtasks["pkid"]);
       mysqli_query($dbconn,"DELETE FROM `tasks` WHERE `pkid` = " . $rowtasks["pkid"]);
